don't set channel values to 0.0 automatically when in color mode other than RGB
The views in color modes R, G and B will only display a pixel's color
information for the regarding channel. The values (not the color
information of the pixel) of the other channels in adverse should stay
as they are: if null then null, if already set to some value through a
slider then that value should be remembered.

Signed-off-by: Stefan Nussbaumer <st9fan@gmail.com>