OSC feedback (almost) working as intended
feedback now responds to feedback being canceled on the client side:
Feedbackstrings are reset to empty every time drawing them to screen has
been finished. However, due to asynchronicity between OscEventListener
and drawing to screen rendering is a bit jittery - not all strings are
necessarily drawn with each frame but most of them.

Signed-off-by: Stefan Nussbaumer <st9fan@gmail.com>